requirements.txt
Django<1.7
South==1.0.2
django-mama-cas==1.0.0
raven==5.0.0
psycopg2
pytest
pytest-django
pytest-cov
pytest-xdist
django-grappelli<2.6.1
python-social-auth
